+++ __tests__/fixtures/loader.js
@@ -1,5 +1,5 @@
-require("./test.gif");
-require("./test.jpg");
-require("./test.png");
-require("./test.svg");
-require("./test.json");
+require("./loader-test.gif");
+require("./loader-test.jpg");
+require("./loader-test.png");
+require("./loader-test.svg");
+require("./loader-test.json");